requirements.txt
opencv-python
apriltag
cmake
pynetworktables
pkgconfig
#robotpy-cscore
tkinter
